portfolio int test, amountEur rouded to 2 decimals, exception handler for EntityNotFound

+++ src/main/java/com/prax/crypto/account/AppUserService.java
@@ -49,6 +49,9 @@ public class AppUserService {
     }
 
     public void delete(Integer id) {
+        if (!appUserRepository.existsById(id)) {
+            throw new EntityNotFoundException("User not found");
+        }
         appUserRepository.deleteById(id);
     }
